+++ homebrew_releaser/app.py
@@ -117,8 +117,6 @@ class App:
                 },
             )
 
-            logger.debug(f'Checksum entry: {archive_checksum_entry}')
-
             Utils.write_file(CHECKSUM_FILE, archive_checksum_entry, 'a')
 
         logger.info(f'Generating Homebrew formula for {GITHUB_REPO}...')